API References

Yunir.AyahEmbeddings — Type

Ayah Embeddings Type

params: numslices - number of slices varslices - variability of slices, parameter for Dirichlet distribution

Yunir.AyahMidpoints — Type

Midpoints Generator

params: n_ayahs - total number of ayahs n - number of samples of slices to generate slicer - Slicer configuration

Yunir.Harakaat — Type

Harakaat(char::Union{String, Char}, is_tanween::Bool)

Create a Harakaat object using char as the Arabic orthography, which is expected to be the short vowels, sukūn, and tanween.

julia> fatha = arabic("a")
julia> Harakaat(fatha, false)
Harakaat("َ", false)

Yunir.Segment — Type

Segment(text::String, harakaat::Array{Harakaat})

Create a Segment object from text, which is the form of the segments of syllables, where vowels of which are also listed as harakaat.

julia> bw_segment = "~aH?Hiy"
julia> Segment(bw_segment, Harakaat[Harakaat("a", false), Harakaat("i", false)])
Segment("~aH?Hiy", Harakaat[Harakaat("a", false), Harakaat("i", false)])

Yunir.Sequence — Type

Sequence(sequence::Vector{String}, y_axis::Vector{Int64})

Create a Sequence object for the rhythmic sequence data, with y_axis as its y-axis ticks for plotting.

Yunir.Slicer — Type

Slicer Configuration

params: numslices - number of slices varslices - variability of slices, parameter for Dirichlet distribution dist_formula - distance metric for measuring similarity or distances of the slices

Yunir.Syllabification — Type

Rhyme(is_quran::Bool, syllable::Syllable)

Create a Rhyme object with specifics for the syllable contructed through Syllable. It also takes argument for is_quran to handle Qur'an input, which does not recite the last vowel in every last word of the verse. The following code creates a Syllable, which specifies 3 syllables with 1 leading and trailing characters to include.

julia> ar_raheem_alamiyn = ["ٱلرَّحِيمِ", "ٱلْعَٰلَمِينَ"]
2-element Vector{String}:
 "ٱلرَّحِيمِ"
 "ٱلْعَٰلَمِينَ"
julia> r = Syllabification(true, Syllable(1, 1, 3))
Rhyme(true, Syllable{Int64}(1, 1, 3))
julia> output = r.(ar_raheem_alamiyn, true)
2-element Vector{Segment}:
 Segment("َّح?حِي", Harakaat[Harakaat("َ", false), Harakaat("ِ", false)])
 Segment("عَٰ?لَم?مِي", Harakaat[Harakaat("َ", false), Harakaat("َ", false), Harakaat("ِ", false)])
julia> encode.(output)
2-element Vector{Segment}:
 Segment("~aH?Hiy", Harakaat[Harakaat("a", false), Harakaat("i", false)])
 Segment("Ea`?lam?miy", Harakaat[Harakaat("a", false), Harakaat("a", false), Harakaat("i", false)])

Yunir.Syllabification — Method

(r::Syllabification)(text::String, isarabic::Bool=false)

Call function for the Syllabification object. It extracts the rhyme features of text using the options from the Syllabification object specified by r. It can handle both Arabic and Buckwalter input by toggling isarabic.

Note: This will only work with @transliterator :default

julia> ar_raheem = "ٱلرَّحِيمِ"
"ٱلرَّحِيمِ"
julia> r = Syllabification(true, Syllable(1, 2, 1))
Syllabification(true, Syllable{Int64}(1, 2, 1))
julia> output = r(ar_raheem, true)
Segment("حِيم", Harakaat[Harakaat("ِ", false)])
julia> encode(output)
Segment("Hiym", Harakaat[Harakaat("i", false)])

Yunir.Syllable — Type

Syllable{T <: Number}(
    lead_nchars::T,
    trail_nchars::T,
    nvowels::T
)

Create a Syllable object specifying the number of vowels nvowels to capture, and also the number of leading (lead_nchars) and trailing characters (trail_nchars) around the vowel. This object is used as an input for the Rhyme object. The following code creates a Syllable, which specifies 3 syllables with 1 leading and trailing characters to include.

julia> Syllable(1, 1, 3)
Syllable{Int64}(1, 1, 3)

Base.join — Function

join(harakaat::Array{Harakaat})

Join function for handling Harakaat object. It joins the harakaat together with ? separator.

julia> ar_raheem = "ٱلرَّحِيمِ"
"ٱلرَّحِيمِ"

julia> r = Rhyme(true, Syllable(1, 2, 2))
Rhyme(true, Syllable{Int64}(1, 2, 2))

julia> output = r(ar_raheem, true)
Segment("َّحِ?حِيم", Harakaat[Harakaat("َ", false), Harakaat("ِ", false)])

julia> join(encode(output).harakaat)
"a?i"

Yunir.align — Method

align(src::Array{String}, tgt::Array{String}; 
	costmodel::CostModel=CostModel(match=0, mismatch=1, insertion=1, deletion=0),
	store_results::Bool=true
)

ALign tgt array of texts to src array of texts using a particular costmodel from BioAlignments.jl. store_results if results of alignment are stored or returned, otherwise, only the scores are returned.

Yunir.align — Method

align(src::String, tgt::String; costmodel::BioAlignments.CostModel=BioAlignments.CostModel(match=0, mismatch=1, insertion=1, deletion=1))

Align tgt string to src string using a particular costmodel from BioAlignments.jl.

Yunir.arabic — Method

arabic(s::String[, encoder::AbstractEncoder])

Encode the String object into Arabic characters. Custom encoder generated from @transliterator can be provided, but default is Buckwalter.

Examples

julia> bw_basmala = "bisomi {ll~ahi {lr~aHoma`ni {lr~aHiymi"
julia> arabic(bw_basmala)
"بِسْمِ ٱللَّهِ ٱلرَّحْمَٰنِ ٱلرَّحِيمِ"

Yunir.clean — Function

clean(s::String, replace_non_ar::String, target_regex::Regex)

Clean the input text by replacing all non-Arabic texts with a string input.

Yunir.crossover! — Method

Parent's Chromosome Crossover

The parents chromosome are crossed-over

Yunir.dediac — Method

dediac(s::String; isarabic::Bool=true)

Dediacritize the input String object.

Examples

julia> bw_basmala = "bisomi {ll~ahi {lr~aHoma`ni {lr~aHiymi"
julia> dediac(bw_basmala)
"bsm {llh {lrHmn {lrHym"
julia> dediac(arabic(bw_basmala))
"بسم ٱلله ٱلرحمن ٱلرحيم"

Yunir.encode — Method

encode(s::String)

Transliterate the input String object using Buckwalter.

Examples

julia> ar_basmala = "بِسْمِ ٱللَّهِ ٱلرَّحْمَٰنِ ٱلرَّحِيمِ"
julia> encode(ar_basmala)
"bisomi {ll~ahi {lr~aHoma`ni {lr~aHiymi"

Yunir.encode — Method

encode(s::Union{Char,String}, encoder::AbstractEncoder)

Transliterate the input String object using a custom encoder. Custom encoder is generated using the @transliterator.

Yunir.fitness — Method

Circular-wise Computation of the Distance of Slices

params: five_nums - five number summaries slicer - Slicer configuration dist - a Distances UnionSemiMetric

Yunir.five_summary — Method

Compute Five Number Summary

five_summary(v::Vector{T}) where T<:Union{Float32,Float64} params: v - data (e.g. embeddings)

five_summary(slices::Vector{Vector{Matrix{T}}}) where T<:Union{Float32,Float64} params: slices - slices of data (e.g. ayah embeddings)

Yunir.gen_slices — Method

Slices Generator

params: ayahs_emb - the embeddings to slice midpoints - the midpoints used for slicing

Yunir.isfeat — Method

isfeat(x::Orthography, y::Type{<:AbstractConsonant})

checks if x is a y feature.

julia> ar_basmala = "بِسْمِ ٱللَّهِ ٱلرَّحْمَٰنِ ٱلرَّحِيمِ";
julia> arb_token = tokenize(ar_basmala)
4-element Vector{String}:
 "بِسْمِ"
 "ٱللَّهِ"
 "ٱلرَّحْمَٰنِ"
 "ٱلرَّحِيمِ"
julia> arb_parsed2 = parse.(Orthography, arb_token)
4-element Vector{Orthography}:
 Orthography(Type[Ba, Kasra, Seen, Sukun, Meem, Kasra])
 Orthography(Type[AlifHamzatWasl, Lam, Lam, Shadda, Fatha, Ha, Kasra])
 Orthography(Type[AlifHamzatWasl, Lam, Ra, Shadda, Fatha, HHa, Sukun, Meem, Fatha, AlifKhanjareeya, Noon, Kasra])
 Orthography(Type[AlifHamzatWasl, Lam, Ra, Shadda, Fatha, HHa, Kasra, Ya, Meem, Kasra])
julia> isfeat(arb_parsed2[1], AbstractLunar)
6-element BitVector:
 1
 0
 0
 0
 1
 0

Yunir.normalize — Function

normalize(s::String)

Normalize a Arabic or Buckwalter String characters.

Examples

julia> normalize("بِسْمِ ٱللَّهِ ٱلرَّحْمَٰنِ ٱلرَّحِيمِ")
"بِسْمِ اللَّهِ الرَّحْمَانِ الرَّحِيمِ"

Yunir.normalize — Method

normalize(s::String, char::Symbol; isarabic::Bool=true)

Normalize a specific Arabic or Buckwalter String character (chars).

Examples

julia> ar_basmala = "بِسْمِ ٱللَّهِ ٱلرَّحْمَٰنِ ٱلرَّحِيمِ"
julia> normalize(ar_basmala, :alif_khanjareeya) === "بِسْمِ ٱللَّهِ ٱلرَّحْمَانِ ٱلرَّحِيمِ"

Yunir.normalize — Method

normalize(s::String, chars::Array{Symbol,1}; isarabic::Bool=true)

Normalize a specific Arabic or Buckwalter String character/s (chars).

Examples

julia> ar_basmala = "بِسْمِ ٱللَّهِ ٱلرَّحْمَٰنِ ٱلرَّحِيمِ"
julia> normalize(ar_basmala, [:alif_khanjareeya, :hamzat_wasl]) === "بِسْمِ اللَّهِ الرَّحْمَانِ الرَّحِيمِ"

Yunir.selection — Method

Tournament Selection of Parent's Chromosome

The function selects the parent's chromosome using tournament selection

params: ayahs_emb - Ayah embeddings mp - Midpoints of the Ayah k - The size of samples from where the parent is chosen

Yunir.sequence — Method

sequence(segments::Array{Segment}, type::Union{Type{Harakaat},Type{Segment}})

Extracts the sequence of the segments by indexing it into xandy, wherexis the index of the segment, andyis the index of its vowels or harakaat. It returns a tuple containing the followingy,y_dict(the mapping dictionary with key represented byxand value represented byy),syllablesrepresented byx`.

julia> ar_raheem_alamiyn = ["ٱلرَّحِيمِ", "ٱلْعَٰلَمِينَ"]
2-element Vector{String}:
 "ٱلرَّحِيمِ"
 "ٱلْعَٰلَمِينَ"

julia> r = Rhyme(true, Syllable(1, 1, 2))
Rhyme(true, Syllable{Int64}(1, 1, 2))

julia> segments = encode.(r.(ar_raheem_alamiyn, true))
2-element Vector{Segment}:
 Segment("~aH?Hiy", Harakaat[Harakaat("a", false), Harakaat("i", false)])
 Segment("lam?miy", Harakaat[Harakaat("a", false), Harakaat("i", false)])

julia> sequence(segments, Segment)
(["~aH?Hiy", "lam?miy"], [1, 2], Dict("~aH?Hiy" => 1, "lam?miy" => 2))

julia> sequence(segments, Harakaat)
(["a?i", "a?i"], [1, 1], Dict("a?i" => 1))

julia> syllables, y_vec, y_dict = transition(segments, Harakaat)
(["a?i", "a?i"], [1, 1], Dict("a?i" => 1))

julia> using Makie

julia> using CairoMakie

julia> f = Figure(resolution=(500, 500));

julia> a1 = Axis(f[1,1], 
           xlabel="Ayah Number",
           ylabel="Last Pronounced Syllable

",
           title="Surah Al-Fatihah Rhythmic Patterns

",
           yticks=(unique(y_vec), unique(syllables)), 
       )
Axis with 1 plots:
 ┗━ Mesh{Tuple{GeometryBasics.Mesh{3, Float32, GeometryBasics.TriangleP{3, Float32, GeometryBasics.PointMeta{3, Float32, Point{3, Float32}, (:normals,), Tuple{Vec{3, Float32}}}}, GeometryBasics.FaceView{GeometryBasics.TriangleP{3, Float32, GeometryBasics.PointMeta{3, Float32, Point{3, Float32}, (:normals,), Tuple{Vec{3, Float32}}}}, GeometryBasics.PointMeta{3, Float32, Point{3, Float32}, (:normals,), Tuple{Vec{3, Float32}}}, GeometryBasics.NgonFace{3, GeometryBasics.OffsetInteger{-1, UInt32}}, StructArrays.StructVector{GeometryBasics.PointMeta{3, Float32, Point{3, Float32}, (:normals,), Tuple{Vec{3, Float32}}}, @NamedTuple{position::Vector{Point{3, Float32}}, normals::Vector{Vec{3, Float32}}}, Int64}, Vector{GeometryBasics.NgonFace{3, GeometryBasics.OffsetInteger{-1, UInt32}}}}}}}

julia> lines!(a1, collect(eachindex(syllables)), y_vec)
Lines{Tuple{Vector{Point{2, Float32}}}}

julia> f

Yunir.syllabic_consistency — Method

syllabic_consistency(segments::Vector{Segment}, syllable_timings::Dict{String,Int64})

Compute syllabicconsistency from a given segments and `syllabletimings. THIS WILL ONLY WORK IF THE VOWEL HAS ONLY 1 TRAIL``julia-repl julia> using Yunir julia> using QuranTree julia> crps, tnzl = load(QuranData()); julia> crpstbl = table(crps) julia> tnzltbl = table(tnzl) julia> bwtexts = verses(tnzltbl[2]) julia> texts = string.(split(bwtexts[1])) julia> r = Syllabification(true, Syllable(1, 0, 5)) julia> segments = Segment[] julia> j = 1 julia> for i in texts if j == 1 push!(segments, r(encode(i), isarabic=false, firstword=true, silentlastvowel=false)) elseif j == length(texts) push!(segments, r(encode(i), isarabic=false, firstword=false, silentlastvowel=true)) else push!(segments, r(encode(i), isarabic=false, firstword=false, silentlastvowel=false)) end j += 1 end julia> tajweedtimings = Dict{String,Int64}( "i" => 1, "a" => 1, "u" => 1, "F" => 1, "N" => 1, "K" => 1, "iy" => 2, "aA" => 2, "uw" => 2, "^" => 4 ) julia> syllabicconsistency(segments, tajweedtimings)

Yunir.tokenize — Function

tokenize(s::String)

tokenizes the string input s by space, and also tokenizes the punctuations.

julia> ar_basmala = "بِسْمِ ٱللَّهِ ٱلرَّحْمَٰنِ ٱلرَّحِيمِ"
julia> tokenize(ar_basmala)
4-element Vector{String}:
 "بِسْمِ"
 "ٱللَّهِ"
 "ٱلرَّحْمَٰنِ"
 "ٱلرَّحِيمِ"

Yunir.@transliterator — Macro

@transliterator(dict, name)

Create a custom transliterator using an input dict (Dict object) with its corresponding name as String object. This will automatically update the transliterator inside all functions like arabic, verses, and encode.

Examples

julia> my_encoder = Dict(
    Symbol(Char(0x0621)) => Symbol('('),
    Symbol(Char(0x0622)) => Symbol('''),
    Symbol(Char(0x0623)) => Symbol('&'),
    Symbol(Char(0x0624)) => Symbol('>'),
    Symbol(Char(0x0625)) => Symbol('}'),
    Symbol(Char(0x0626)) => Symbol('<'),
    Symbol(Char(0x0627)) => Symbol('b'),
    Symbol(Char(0x0628)) => Symbol('A'),
    Symbol(Char(0x0629)) => Symbol('t'),
    Symbol(Char(0x062A)) => Symbol('p'),
    Symbol(Char(0x062B)) => Symbol('j'),
    Symbol(Char(0x062C)) => Symbol('v'),
    Symbol(Char(0x062D)) => Symbol('x'),
    Symbol(Char(0x062E)) => Symbol('H'),
    Symbol(Char(0x062F)) => Symbol('*'),
    Symbol(Char(0x0630)) => Symbol('d'),
    Symbol(Char(0x0631)) => Symbol('z'),
    Symbol(Char(0x0632)) => Symbol('r'),
    Symbol(Char(0x0633)) => Symbol('$'),
    Symbol(Char(0x0634)) => Symbol('s'),
    Symbol(Char(0x0635)) => Symbol('D'),
    Symbol(Char(0x0636)) => Symbol('S'),
    Symbol(Char(0x0637)) => Symbol('Z'),
    Symbol(Char(0x0638)) => Symbol('T'),
    Symbol(Char(0x0639)) => Symbol('g'),
    Symbol(Char(0x063A)) => Symbol('E'),
    Symbol(Char(0x0640)) => Symbol('f'),
    Symbol(Char(0x0641)) => Symbol('_'),
    Symbol(Char(0x0642)) => Symbol('k'),
    Symbol(Char(0x0643)) => Symbol('q'),
    Symbol(Char(0x0644)) => Symbol('m'),
    Symbol(Char(0x0645)) => Symbol('l'),
    Symbol(Char(0x0646)) => Symbol('h'),
    Symbol(Char(0x0647)) => Symbol('n'),
    Symbol(Char(0x0648)) => Symbol('Y'),
    Symbol(Char(0x0649)) => Symbol('w'),
    Symbol(Char(0x064A)) => Symbol('F'),
    Symbol(Char(0x064B)) => Symbol('y'),
    Symbol(Char(0x064C)) => Symbol('K'),
    Symbol(Char(0x064D)) => Symbol('N'),
    Symbol(Char(0x064E)) => Symbol('u'),
    Symbol(Char(0x064F)) => Symbol('a'),
    Symbol(Char(0x0650)) => Symbol('~'),
    Symbol(Char(0x0651)) => Symbol('i'),
    Symbol(Char(0x0652)) => Symbol('^'),
    Symbol(Char(0x0653)) => Symbol('o'),
    Symbol(Char(0x0654)) => Symbol('`'),
    Symbol(Char(0x0670)) => Symbol('#'),
    Symbol(Char(0x0671)) => Symbol(':'),
    Symbol(Char(0x06DC)) => Symbol('{'),
    Symbol(Char(0x06DF)) => Symbol('"'),
    Symbol(Char(0x06E0)) => Symbol('@'),
    Symbol(Char(0x06E2)) => Symbol(';'),
    Symbol(Char(0x06E3)) => Symbol('['),
    Symbol(Char(0x06E5)) => Symbol('.'),
    Symbol(Char(0x06E6)) => Symbol(','),
    Symbol(Char(0x06E8)) => Symbol('-'),
    Symbol(Char(0x06EA)) => Symbol('!'),
    Symbol(Char(0x06EB)) => Symbol('%'),
    Symbol(Char(0x06EC)) => Symbol('+'),
    Symbol(Char(0x06ED)) => Symbol(']')
);
julia> @transliterator my_encoder "MyEncoder"
julia> encode(ar_basmala)
"A~$^l~ :mmiun~ :mziux^lu#h~ :mziux~Fl~"

Yunir.@transliterator — Macro

@transliterator(symbl)

Fallback to the default Buckwalter transliterator.

julia> @transliterator :default
julia> ar_basmala = "بِسْمِ ٱللَّهِ ٱلرَّحْمَٰنِ ٱلرَّحِيمِ"
julia> encode(ar_basmala)
"bisomi {ll~ahi {lr~aHoma`ni {lr~aHiymi"